Fix path name resolution for overlapping build contexts
An incorrect assumption was made in the code that handled build contexts
that were shared, in that the context would be the same. This didn't
take into account build contexts which are a subdirectory of another
build context, which would cause the name resolution to be broken.

Also added a test case to catch this, and slightly changed some helper
utilities in the tests.

Change-type: patch
Connects-to: #27
Signed-off-by: Cameron Diver <cameron@resin.io>