Propagate target platform to builds
This ensures that the build resolves the correct platform for base
images it fetches from the registry.
Includes unit tests to check the platform translation and the actual
build execution.

This requires running a docker daemon >=18.09.3 and experimental
features enabled, to support this tests will only run if a daemon
matching this requirement is available.

Change-type: minor
Signed-off-by: Robert Günzler <robertg@balena.io>